+++ src/App.tsx
@@ -14,6 +14,7 @@ import InicioPage from './pages/inicio'
 import LogisticaPage from './pages/logistica'
 import RepartosPage from './pages/logistica/repartos'
 import ChoferRepartosPage from './pages/logistica/repartos/chofer'
+import PickingPage from './pages/logistica/picking'
 import FinanzasPage from './pages/finanzas'
 import CobrosPage from './pages/cobros'
 import PedidosPage from './pages/pedidos'
@@ -146,6 +147,14 @@ function AppRoutes() {
             element={
               <ModuleRoute moduleKey="logistics.dispatches">
                 <LogisticaPage />
+              </ModuleRoute>
+            }
+          />
+          <Route
+            path="logistica/picking"
+            element={
+              <ModuleRoute moduleKey="logistics.picking">
+                <PickingPage />
               </ModuleRoute>
             }
           />